rename package too
