+++ v1_21/src/main/java/net/countercraft/movecraft/compat/v1_21_1/NextTickProvider.java
@@ -1,4 +1,4 @@
-package net.countercraft.movecraft.compat.v1_21;
+package net.countercraft.movecraft.compat.v1_21_1;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
