+++ v1_21_1/src/main/java/net/countercraft/movecraft/support/v1_21_1/ISmoothTeleport.java
@@ -1,4 +1,4 @@
-package net.countercraft.movecraft.support.v1_21;
+package net.countercraft.movecraft.support.v1_21_1;
 
 import io.papermc.paper.entity.TeleportFlag;
 import net.countercraft.movecraft.SmoothTeleport;
